extract data models to separate folder
