+++ src/app/_models/city.ts
@@ -1,0 +1,15 @@
+export class City {
+    cityId: string;
+    name: string;
+    country: string;
+    province: string;
+    cityInfo: {
+        cityInfoId: number,
+        population: number,
+        citySize: number
+    }
+
+    toString() {
+        return this.name || "";
+    }
+}
