+++ src/app/_models/index.ts
@@ -1,0 +1,9 @@
+import { City } from "./city";
+import { Route } from "./route";
+import { User } from "./user";
+
+export {
+    City,
+    Route,
+    User
+}
