+++ src/app/_models/user.ts
@@ -1,0 +1,11 @@
+export class User {
+    id: number;
+    gender: string;
+    firstName: string;
+    lastName: string;
+    login: string;
+    password: string;
+    email: string;
+    token: string;
+    phone: string;
+}
