+++ src/app/login/login.component.ts
@@ -2,7 +2,8 @@ import { Component, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { AuthService } from '../_services/auth.service';
 import { Router, ActivatedRoute } from '@angular/router';
-import { User } from '../_services/user.service';
+import { User } from "../_models";
+
 
 @Component({
   selector: 'app-login',
