+++ src/app/nav/nav.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from '../_services/auth.service';
 import { Router } from '@angular/router';
-import { User } from '../_services/user.service';
+import { User } from "../_models";
 
 @Component({
   selector: 'app-nav',
