+++ src/app/register/register.component.ts
@@ -1,7 +1,8 @@
 import { Component, ViewChild, OnInit } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
-import { UserService, User } from '../_services/user.service';
+import { UserService } from '../_services/user.service';
+import { User } from "../_models";
 import { AuthService } from '../_services/auth.service';
 
 @Component({
